scope indicator helper to analysis cached
